requirements.txt
fastapi
httpx
pytest
sqlalchemy
databases
pymysql
slowapi
uvicorn
aiosqlite
pyjwt
passlib
password-strength
email-validator
better-profanity
alt-profanity-check
python-multipart
postmarker
aiomysql
sqlmodel